fix(ui): address review findings in agent catalog modals
- Extract checkIdValid/validateField to shared validation.utils.ts
- Merge doCreateAgent/doImportAgent into single doSaveAgent function
- Fix PleaseWaitModal showing wrong message for import operations
- Fix double JSON.parse in VersionPage getAgentCardFromContent()
- Add props.agentCard to EditAgentModal useEffect dependency array
