+++ ui/ui-app/src/app/pages/agents/components/modals/CreateAgentModal.tsx
@@ -18,37 +18,11 @@ import { Modal } from "@patternfly/react-core/deprecated";
 import { ExclamationCircleIcon } from "@patternfly/react-icons";
 import { AgentCard, AgentCardEditor } from "@app/components/agentCard";
 import { CreateArtifact } from "@sdk/lib/generated-client/models";
-
-type ValidType = "default" | "success" | "error";
+import { checkIdValid, validateField, ValidType } from "@utils/validation.utils.ts";
 
 type Validities = {
     groupId?: ValidType;
     artifactId?: ValidType;
-};
-
-const checkIdValid = (id: string | undefined | null): boolean => {
-    if (!id) {
-        return true;
-    }
-    const isAscii = (str: string) => {
-        for (let i = 0; i < str.length; i++) {
-            if (str.charCodeAt(i) > 127) {
-                return false;
-            }
-        }
-        return true;
-    };
-    return id.indexOf("%") === -1 && isAscii(id);
-};
-
-const validateField = (value: string | undefined | null): ValidType => {
-    if (!checkIdValid(value)) {
-        return "error";
-    }
-    if (value === undefined || value === null || value === "") {
-        return "default";
-    }
-    return "success";
 };
 
 const EMPTY_AGENT_CARD: AgentCard = {
