+++ ui/ui-app/src/app/pages/agents/components/modals/ImportAgentModal.tsx
@@ -20,33 +20,7 @@ import { UrlUpload } from "@apicurio/common-ui-components";
 import { AgentCard, AgentCardViewer } from "@app/components/agentCard";
 import { UrlService, useUrlService } from "@services/useUrlService.ts";
 import { CreateArtifact } from "@sdk/lib/generated-client/models";
-
-type ValidType = "default" | "success" | "error";
-
-const checkIdValid = (id: string | undefined | null): boolean => {
-    if (!id) {
-        return true;
-    }
-    const isAscii = (str: string) => {
-        for (let i = 0; i < str.length; i++) {
-            if (str.charCodeAt(i) > 127) {
-                return false;
-            }
-        }
-        return true;
-    };
-    return id.indexOf("%") === -1 && isAscii(id);
-};
-
-const validateField = (value: string | undefined | null): ValidType => {
-    if (!checkIdValid(value)) {
-        return "error";
-    }
-    if (value === undefined || value === null || value === "") {
-        return "default";
-    }
-    return "success";
-};
+import { checkIdValid, validateField, ValidType } from "@utils/validation.utils.ts";
 
 export type ImportAgentModalProps = {
     isOpen: boolean;
