+++ ui/ui-app/src/utils/index.ts
@@ -3,3 +3,4 @@ export * from "./labels.utils";
 export * from "./object.utils";
 export * from "./rest.utils";
 export * from "./string.utils";
+export * from "./validation.utils";
